+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 _Changes in the next release_
 
+### Added
+- New commands: mute toggle, play, pause, stop ([#34](https://github.com/unfoldedcircle/integration-appletv/issues/34)).
+
 ---
 
 ## v0.16.0 - 2025-03-24